TextFields turned into multilines. Moved new row hotkey.
TextFields can now span multiple lines making the contents possibly more
readable and enabling support for multiline table fields.
To achieve that adding a new row to the table has been moved to pressing
"SHIFT + ENTER" instead of "ENTER".
As of now adding completely blank lines to a textfield will break the
syntax when pasting the export to a textile editor.